Initial CG SVM support for PoCL-R
Works only when mmap start addresses can be made to match.

Rework the internal SVM implementation to use a cl_mem shadow buffer.
The shadow buffer will be then passed to standard cl_mem APIs for
mapping, unmapping, memcpy, migrate etc.

This unifies all the automated migration, content size buffer, etc.
handling code to avoid special casing for SVM buffers.

Converted PoCL-D/SVM to use SubBuffers in the server-side for managing
the chunks: Using parent Buffers doesn't work as a wrapper since the
cl_mems are expected to point to the beginning of the SVM buffer,
which leads to undefined results when wrapping chunks from the large
SVM allocation.

With this, chipStar 1.1-RC1 can run at least the simple_kernel and
MatrixMultiply correctly (when "the mmoon" happens to land in the
right phase).

chipStar 1.1-RC3 added the runtime linking for enhanced portability
feature, which requires cl{Compile,Link}Program support which the
remote driver currently lacks.

+++ pocld/common.hh
@@ -400,8 +400,11 @@ struct EventPair {
 
 std::string hexstr(const std::string &);
 
-// Type for the buffer allocation identifier.
-typedef uint32_t BufferId_t;
+// Type for the buffer allocation identifier, a running number starting from
+// 0. Note: In some parts functions, the input can be declared as a buffer
+// id which might be treated as a raw 64b SVM pointer instead of a cl_mem
+// buffer identifier, in case referring to SVM buffers.
+typedef uint64_t BufferId_t;
 
 #ifdef __GNUC__
 #pragma GCC visibility pop
